apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
redis==5.2.1
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
alembic==1.14.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
pytest==8.3.2
httpx==0.27.0


